Homework Done: It can be found in Leason3. Also organized the code a bit.
