+++ Alexandru_Plesa/Lesson3/ConsoleApp3/ConsoleApp3.csproj
@@ -43,8 +43,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="GeometryFactory.cs" />
+    <Compile Include="iGeometryFactory.cs" />
+    <Compile Include="iStudentRepository.cs" />
+    <Compile Include="Point2D.cs" />
+    <Compile Include="Point3D.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Student.cs" />
+    <Compile Include="StudentRepository.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
