+++ Alexandru_Plesa/Lesson3/ConsoleApp3/Student.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConsoleApp3
+{
+    public class Student
+    {
+        public string Name;
+        public int ID;
+        public DateTime DateOfBirth;
+    }
+}
